Authentication and authorization using jwt added

+++ controllers/api/v1/posts_api.js
@@ -34,7 +34,7 @@ module.exports.destroy = async function (req, res) {
         let post = await Post.findById(req.params.id);
         // authorize , only perosn posted that post should be able to del
         //.id means converting the object id into string
-        // if (post.user == req.user.id) {
+        if (post.user == req.user.id) {
 
             post.remove();
             await Comment.deleteMany({ post: req.params.id });
@@ -45,11 +45,12 @@ module.exports.destroy = async function (req, res) {
                 message : "Post and associated comments deleted successfully!"
             })
         
-        // } else {
-
-        //     req.flash('error' , 'You cannot delete this post !');
-        //     return res.redirect('back');
-        // }
+        } else {
+            return res.json(401,{
+                message : "You cannot delete this post !"
+            });
+            
+        }
     } catch (err) {
         console.log(err);
         return res.json(500,{
